+++ docs/guides/connecting_to_your_data/database/postgres.md
@@ -7,7 +7,7 @@ import ConnectionStringDetails from '../reference/configuration_explanation.md'
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 
-This guide will help you connect Great Expectations to data in a Postgresql database. We will be building a `Datasource` configuration, and using it to verify connectivity. By the end of this how-to guide, you should have a `Batch` representing a slice of data in your database, and a working code snippet that can be used as a starting point for more complex configurations.
+This guide will help you connect Great Expectations to data in a Postgresql database. You will be building a `Datasource` configuration, and using it to verify connectivity. By the end of this how-to guide, you should have a `Batch` representing a slice of data in your database, and a working code snippet that can be used as a starting point for more complex configurations.
 
 :::note Prerequisites: This how-to-guide assumes you have already:
 - Completed the [Getting Started Tutorial.](../../tutorials/quick-start.md)
@@ -17,7 +17,15 @@ This guide will help you connect Great Expectations to data in a Postgresql data
 
 ## Steps
 
-### 1. Determine how to add credentials to configuration
+### 1. Install required packages
+
+Next make sure you have the necessary packages for Great Expectations to connect to your postgres database.
+
+```console
+pip install sqlalchemy psycopg2
+```
+
+### 2. Determine how to add credentials to configuration
 
 Great Expectations provides multiple methods of providing credentials for accessing databases. For our current example will use a `connection_string`, but other options include providing an Environment Variable, and loading from a Cloud Secret Store.  
 
@@ -33,13 +41,6 @@ postgresql+psycopg2://postgres:@localhost/test_ci
 </p>
 </details>
 
-### 2. Install required packages
-
-Next make sure you have the necessary packages for Great Expectations to connect to your postgres database.
-
-```console
-pip install sqlalchemy psycopg2
-```
 
 ### 3. Load the DataContext into memory
 
@@ -55,14 +56,14 @@ Open up a Jupyter Notebook in the same directory as the `great_expectations/` fo
 
 Load your DataContext into memory using the `get_context()` method.
 
-```python file=../../../../integration/code/connecting_to_your_data/database/postgres.py#L13
+```python file=../../../../integration/code/connecting_to_your_data/database/postgres.py#L16
 ```
 
 ### 4. Write your configuration as a YAML
 
 Here is an example configuration:
 
-```python file=../../../../integration/code/connecting_to_your_data/database/postgres.py#L19-L33
+```python file=../../../../integration/code/connecting_to_your_data/database/postgres.py#L19-33
 ```
 
 :::note What does the configuration contain?
@@ -87,7 +88,7 @@ The configuration also contains 2 `DataConnectors` by default:
 
 Save the configuration into your DataContext by running the `add_datasource()` function.
 
-```python file=../../../../integration/code/connecting_to_your_data/database/postgres.py#L37
+```python file=../../../../integration/code/connecting_to_your_data/database/postgres.py#L36
 ```
 
 ### 6. Test Configuration
@@ -97,15 +98,15 @@ Test your configuration by retrieving a Validator from Great Expectations using 
 <Tabs
   defaultValue='runtime_batch_request'
   values={[
-  {label: 'Using a Query', value:'runtime_batch_request'},
-  {label: 'Naming a Table', value:'batch_request'},
+  {label: 'Using a query', value:'runtime_batch_request'},
+  {label: 'Using a table name', value:'batch_request'},
   ]}>
   <TabItem value="runtime_batch_request">
 
-Here is an example of loading a batch from a query. As you can see in the following snippet, when we fetch a Batch of data, we actually create a Validator, which is a Batch + ExpectationSuite. This allows you to perform operations like `.head()` to see 
+Here is an example of loading a batch from a query. As you can see in the following snippet, when you fetch a Batch of data, you actually create a Validator, which is a Batch + ExpectationSuite. This allows you to perform operations like `.head()` to see 
 the first few rows of your table, as well as run Expectations directly. 
 
-```python file=../../../../integration/code/connecting_to_your_data/database/postgres.py#L41-L51
+```python file=../../../../integration/code/connecting_to_your_data/database/postgres.py#L39-L53
 ```
 
 :::note What does the RuntimeBatchRequest contain?
@@ -116,21 +117,20 @@ the first few rows of your table, as well as run Expectations directly.
 :::
 
 
-
   </TabItem>
 
   <TabItem value="batch_request">
 
-Here is an example of loading a batch by naming a table. As you can see in the following snippet, when we fetch a Batch of data, we actually create a Validator, which is a Batch + ExpectationSuite. This allows you to perform operations like `.head()` to see 
+Here is an example of loading a batch by naming a table. As you can see in the following snippet, when you fetch a Batch of data, you actually create a Validator, which is a Batch + ExpectationSuite. This allows you to perform operations like `.head()` to see 
 the first few rows of your table, as well as run Expectations directly. 
 
-```python file=../../../../integration/code/connecting_to_your_data/database/postgres.py#L54-L62
+```python file=../../../../integration/code/connecting_to_your_data/database/postgres.py#L56-L67
 ```
 
 :::note What does the BatchRequest contain?
 1. `datasource_name` and `data_connector_name` are directly from our `Datasource` configuration.
-2.  `data_asset_name` is `taxi_data`, which is the name of the table we want to retrieve as a batch.  
-The reason we can do this is because of the `InferredAssetDataConnector` that is configured to retrieve `whole_table` by default.
+2.  `data_asset_name` is `taxi_data`, which is the name of the table you want to retrieve as a batch.  
+The reason you can do this is because of the `InferredAssetDataConnector` that is configured to retrieve `whole_table` by default.
 
 **To Discuss**: How much do we mention `ActiveDataConnector` here?
 :::
@@ -144,3 +144,9 @@ The reason we can do this is because of the `InferredAssetDataConnector` that is
 ## Additional Notes
 
 <AddingCredentials />
+
+
+### Here is the full script from which snippets were taken
+
+```python file=../../../../integration/code/connecting_to_your_data/database/postgres.py#L1-L67
+```
