+++ integration/code/connecting_to_your_data/database/postgres.py
@@ -2,19 +2,18 @@ from ruamel import yaml
 
 import great_expectations as ge
 from great_expectations.core.batch import Batch
-from great_expectations.util import load_data_into_database
+from integration.code.connecting_to_your_data.database.util import (
+    load_data_into_database,
+)
 
 CONNECTION_STRING = """postgresql+psycopg2://postgres:@localhost/test_ci"""
 load_data_into_database(
     "taxi_data",
-    "../../../../integration/fixtures/data/reports/yellow_tripdata_sample_2019-01.csv",
+    "./data/reports/yellow_tripdata_sample_2019-01.csv",
     CONNECTION_STRING,
 )
 
-# context = ge.get_context()
-context = ge.DataContext(
-    context_root_dir="../../../../integration/fixtures/runtime_data_taxi_monthly/great_expectations"
-)
+context = ge.get_context()
 
 # with a connection_string
 config = """
@@ -34,10 +33,8 @@ data_connectors:
 """
 config = config.replace("<YOUR_CONNECTION_STRING_HERE>", CONNECTION_STRING)
 
-print(f"Adding Datasource")
 context.add_datasource(**yaml.load(config))
 
-print(f"RuntimeBatchRequest")
 # First test for RuntimeBatchRequest using a query
 batch_request = ge.core.batch.RuntimeBatchRequest(
     datasource_name="my_postgres_datasource",
