+++ tests/integration/test_script_runner.py
@@ -26,22 +26,23 @@ integration_test_matrix = [
         "data_context_dir": "integration/fixtures/runtime_data_taxi_monthly/great_expectations",
         "data_dir": "integration/fixtures/data",
     },
-    # {
-    #     "name": "postgres_runtime_golden_path",
-    #     "base_dir": file_relative_path(__file__, "../../"),
-    #     "data_context_dir": "integration/fixtures/runtime_data_taxi_monthly/great_expectations",
-    #     "user_flow_script": "integration/code/query_postgres_runtime_data_connector.py",
-    #     "extra_backend_dependencies": BackendDependencies.POSTGRESQL,
-    # },
     {
-        "name": "pandas_two_batch_requests_two_validators",
+        "name": "postgres_runtime_golden_path",
+        "data_dir": "integration/fixtures/data",
         "base_dir": file_relative_path(__file__, "../../"),
-        "data_context_dir": "tests/integration/fixtures/yellow_trip_data_pandas_fixture/great_expectations",
-        "data_dir": "tests/test_sets/taxi_yellow_trip_data_samples",
-        "user_flow_script": "tests/integration/fixtures/yellow_trip_data_pandas_fixture/two_batch_requests_two_validators.py",
-        "expected_stderrs": "",
-        "expected_stdouts": "",
+        "data_context_dir": "integration/fixtures/runtime_data_taxi_monthly/great_expectations",
+        "user_flow_script": "integration/code/connecting_to_your_data/database/postgres.py",
+        "extra_backend_dependencies": BackendDependencies.POSTGRESQL,
     },
+    # {
+    #     "name": "pandas_two_batch_requests_two_validators",
+    #     "base_dir": file_relative_path(__file__, "../../"),
+    #     "data_context_dir": "tests/integration/fixtures/yellow_trip_data_pandas_fixture/great_expectations",
+    #     "data_dir": "tests/test_sets/taxi_yellow_trip_data_samples",
+    #     "user_flow_script": "tests/integration/fixtures/yellow_trip_data_pandas_fixture/two_batch_requests_two_validators.py",
+    #     "expected_stderrs": "",
+    #     "expected_stdouts": "",
+    # },
 ]
